setup: restructure app
