+++ src/app/rootReducer.ts
@@ -1,0 +1,11 @@
+import { combineReducers } from '@reduxjs/toolkit'
+
+import authReducer from '../features/auth/authSlice'
+import petReducer from '../features/pet/petSlice'
+
+const rootReducer = combineReducers({
+  auth: authReducer,
+  pet: petReducer,
+})
+
+export default rootReducer
