+++ src/app/store.ts
@@ -1,16 +1,19 @@
 import { configureStore } from '@reduxjs/toolkit'
+import type { TypedUseSelectorHook } from 'react-redux'
+import { useDispatch, useSelector } from 'react-redux'
 
-import authReducer from '../features/auth/authSlice'
-import petReducer from '../features/pet/petSlice'
+import rootReducer from './rootReducer'
 
 export const store = configureStore({
-  reducer: {
-    auth: authReducer,
-    pet: petReducer,
-  },
+  reducer: rootReducer,
+  devTools: true,
 })
 
 // Infer the `RootState` and `AppDispatch` types from the store itself
 export type RootState = ReturnType<typeof store.getState>
 // Inferred type: {posts: PostsState, comments: CommentsState, users: UsersState}
 export type AppDispatch = typeof store.dispatch
+
+// Use throughout your app instead of plain `useDispatch` and `useSelector`
+export const useAppDispatch: () => AppDispatch = useDispatch
+export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector
