+++ src/features/auth/authSlice.tsx
@@ -2,8 +2,8 @@ import type { PayloadAction } from '@reduxjs/toolkit'
 import { createSlice, createAsyncThunk } from '@reduxjs/toolkit'
 
 import type { RootState } from '../../app/store'
-import { Status } from '../../constants/enums'
-import type { HttpError } from '../../constants/types'
+import { Status } from '../../common/types/enums'
+import type { HttpError } from '../../common/types/types'
 
 // Define a type for the slice state
 interface Auth {
